Stargazer 2D Platformer Game
Stargazer is a 2D video game that was inspired by those such as Celeste and Super Mario. Unlike other 2D platformers, Stargazer incorporates a weather changing system that will change the appearance and flow of the game, based on your current location's weather. Along with that, Stargazer uses ray tracing for collision detection, and physics equations in the code for movement.